Put render elements to separate folder
