+++ patterns/composite/products_and_boxes/diagram/convert_product_to_render_element.dart
@@ -2,8 +2,9 @@ import 'package:design_patterns_dart/text_canvas.dart';
 
 import '../products/product.dart';
 import '../products/product_leaf.dart';
-import 'render_element.dart';
-import 'render_text.dart';
+import '../render_elements/render_element.dart';
+import '../render_elements/render_text.dart';
+
 
 extension ConvertProductToDiagram on Product {
   RenderElement toRenderElement() {
