+++ patterns/composite/products_and_boxes/render_elements/render_column.dart
@@ -1,5 +1,5 @@
-import 'render_layout.dart';
 import 'render_element.dart';
+import 'render_layout.dart';
 import 'render_position.dart';
 
 class RenderColumn extends RenderLayout {
